+++ pytket/binders/include/UnitRegister.hpp
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _TKET_UnitRegister_H_
-#define _TKET_UnitRegister_H_
+#pragma once
 
 #include "Utils/UnitID.hpp"
 
@@ -62,4 +61,3 @@ class UnitRegister {
 typedef UnitRegister<Bit> BitRegister;
 typedef UnitRegister<Qubit> QubitRegister;
 }  // namespace tket
-#endif  //_TKET_UnitRegister_H_