+++ pytket/pytket/__init__.py
@@ -14,7 +14,6 @@
 
 """Python Interface to CQC tket
 """
-
 from pytket.circuit import (  # type: ignore
     Circuit,
     OpType,
@@ -24,5 +23,7 @@ from pytket.circuit import (  # type: ignore
 import pytket.routing
 import pytket.transform
 import pytket.telemetry
+
+from pytket._version import __version__
 
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore