+++ tket/src/Clifford/CliffTableau.hpp
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _TKET_CliffTableau_H_
-#define _TKET_CliffTableau_H_
+#pragma once
 
 #include "OpType/OpType.hpp"
 #include "Utils/BiMapHeaders.hpp"
@@ -207,5 +206,3 @@ class CliffTableau {
 std::ostream &operator<<(std::ostream &os, CliffTableau const &tab);
 
 }  // namespace tket
-
-#endif  // CLIFFTABLEAU_H_