+++ tket/src/Diagonalisation/DiagUtils.hpp
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _TKET_DiagUtils_H_
-#define _TKET_DiagUtils_H_
+#pragma once
 
 #include "PauliGraph/PauliGraph.hpp"
 
@@ -35,9 +34,4 @@ typedef std::map<QubitPauliTensor, Expr, cmp_tensors> QubitOperator;
 void insert_into_gadget_map(
     QubitOperator &gadget_map, const PauliGadgetProperties &pgp);
 
-void insert_into_gadget_map(
-    QubitOperator &gadget_map, const std::pair<QubitPauliTensor, Expr> &pgp);
-
 }  // namespace tket
-
-#endif  // _TKET_DiagUtils_H_