+++ tket/src/Gate/Gate.hpp
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _TKET_Ops_Gate_H_
-#define _TKET_Ops_Gate_H_
+#pragma once
 
 #include "Ops/Op.hpp"
 #include "Utils/Json.hpp"
@@ -83,5 +82,3 @@ class Gate : public Op {
 };
 
 }  // namespace tket
-
-#endif