+++ tket/src/Ops/MetaOp.hpp
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _TKET_Ops_MetaOp_H_
-#define _TKET_Ops_MetaOp_H_
+#pragma once
 
 #include "Op.hpp"
 #include "Utils/Json.hpp"
@@ -53,5 +52,3 @@ class MetaOp : public Op {
 };
 
 }  // namespace tket
-
-#endif