+++ tket/tests/Gate/GatesData.hpp
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _TKET_TESTS_Gate_GatesData_H_
-#define _TKET_TESTS_Gate_GatesData_H_
+#pragma once
 
 #include <map>
 #include <vector>
@@ -42,5 +41,3 @@ struct GatesData {
 
 }  // namespace internal
 }  // namespace tket
-
-#endif